debug: fix hover variable value disappears on key down
Fixes #226474

It looks like this bug is at least 9 years old! It took me a minute to figure out what the intention was here.

I think the goal was to keep showing the hover if the user is pressing ctrl/cmd without sticky hovers enabled (e.g. to allow copying) but it was in the wrong check and would confusingly dismiss the hover when the mouse moved if ctrl/cmd was pressed.