Try to simplify usage of `createAndFillInContextMenuActions`  (#232546)
Try to simplify `createAndFillInContextMenuActions`

This is one of the rare functions that we have that uses an out parameter. This often results in a lot of boilerplate code to call it

In this PR I've renamed it to `getContextMenuActions` to make use of a standard return type. Instead of an overload, there are now two different versions of the function: `getContextMenuActions` and `getFlatContextMenuActions`