Fix toggling wrapping sometimes losing track of DOM nodes
On flush, the DOM renderer expects the following render to have lines
and as such defers clearing the innerHTML of the line collection in
order to save a little time. Since this GPU renderer may not actually
render lines however, this ends up having the possibility to lose track
of rendered DOM lines and they would stick around until the next flush
where it actually does clear the parent out.

This was mostly reproducible when toggling wrapping in markdown files,
but it was somewhat difficult I believe depending on how many lines
would pass canRender or not which changes at various stages of view
events being fired.

Since this is a relatively risky and unnecessary change to apply when
GPU acceleration is off, it will only change behavior when useGpu is
true.

Fixes #237527