mcp: button/command for adding MCP servers (#243675)
An editor content button is shown on mcp.json to add a server.
Additionally, the command is accessible on the command palette and when
invoked from there will prompt the user to pick where to save their
config.

With registry support I envision the "Select Server Type" would list
servers in addition to letting users provide their own.

Closes #243617