tools: allow saving trust for MCP tools
This adds an optional `allowAutoConfirm`, set for MCP tools, to the
`IToolConfirmationMessages`. When preset users have the option to allow,
deny, always allow in workspace, or always allow the tool to run
globally.

This closes #243620 with a better solution imo, because it doesn't
require extra manual configuration from the user and can is done at a
per-tool level. While we could have some tool-to-confirmation map
in mcp.json, I think this much easier to deal with.

The `Chat: Reset Tool Confirmations` can be used to forget trusted
too preferences.

cc @jrieken, let me know if you want me to clean up anything here