mcp: fix issues with bat/cmd based mcp commands
Applies the necessary handling post Node.js CVE-2024-27980 to spawn
bat/cmd processes.

Some of these (npx in fnm for example) are based on the directory
they're run in, which seem to require referring to them by their
absolute path. Do so. And update `findExecutable` to correctly look for
PATHEXT's on Windows.