Initial support of OAuth in MCP (#249056)
This hooks up everything auth-y to MCP. It's basically implemented by following what we do in Auth land for extensions and doing that again for MCP servers... so it's a lot of code, but it's logic is the same for extensions for the most part.

The MCP side of things... this follows the spec:
https://modelcontextprotocol.io/specification/2025-03-26/basic/authorization

To discover the server metadata needed to get a token to add to MCP requests.

This works by leveraging the new Auth issuers API and mapping issuers to auth providers.

What will come in the next PR is dynamic creation of an auth provider based on this server information.