vscode api: Raise compatibility to webview content
The webview (or its  `acquireVsCodeApi` injector)
adds the vscode api but also deletes the properties
`window.parent`, `window.top` and `window.frameElement`.

This voilates the HTML Spec and makes problems with
some code like `if (window.parent === window)` to
detect if we are running in an iframe.

Yes, this affects Hyrum's Law and https://xkcd.com/1172/
but detecting `acquireVsCodeApi` object itself feels to be cleaner anyway.