Avoid re-sanitizing markdown during rendering
Currently rendering markdown sanitizes the content twice. When creating the output element it also goes: `string -> dom-> string -> dom` (with other hops potentially in the implementation)

This change changes to a single pass that goes `string -> dom` with a single sanitize call. Since we are moving away from DomParser, the output is slightly changed