api: allow byok model providers to hint at good edit tools (#268506)
* api: allow byok model providers to hint at good edit tools

For our own models, we set up the edit tools they get. For generic open-
ended providers like OpenRouter, we will have a 'learning' functionality.
But for BYOK providers that can provide a curated set of models (e.g.
Cerebras) we should let them tell us what edit tool is appropriate for
models they give us so users have a good experience from request 0.

The API is intentionally loose because edit tools continue to evolve and
our set of available tools will continue to change. So I do not surface
any of these in actual API types, but give a string array of hints the
provider can give us. If in the future none of the tools they hint are
ones we recognize, as the space evolves, then the chat extension will
fall back to its default learning behavior.

* comment