Use setTimeout0 to defer setting Count til element is fully shown for now (#270825)
This is a workaround for now for an accessibility bug that should be addressed.

The way that the checked count works is that it's a `CountBadge` widget wrapped in an `aria-live` that reports when the text content (aka the number) changes.

However, when the QuickPick is initially shown to the user, the resolved count is done so before actually made visible... `aria-live` does not read out anything that is changed when hidden.

I think a proper change here would be to have a separate `aria-live` element that is always off screen (so "technically" visible)... that can be updated when a Count is made visible... but this small fix is fine to at least give screen reader users a count.

Fixes https://github.com/microsoft/vscode/issues/258617