chat - fix height compute of tree (#270876)
The `gap: 4px` directive for the chat input part needs to be accounted for.

In addition, the tree container itself should not require to get a height applied provided its parent list container has a proper height.

This fixes issues where the blue outline box would be cut off at the bottom.