edits: integration point for externally-editing agent clis
Allows external agent/CLI edits to be harnessed in our edit session
implementation. There is a callback API to make an external edit. While
the callback is running, we assume that all edits made were done by the
agent and update our internal models based on a before/after snapshot.