Disable cleanup contribution (#275700)
Fixes https://github.com/microsoft/vscode/issues/272359

This contribution is causing servers that are allowed to be marked as not allowed because the servers are not available at startup. We shouldn't rely on servers being available at startup...

This needs to be re-done to clean up correctly, but for now let's just disable it.

This means that approvals will accumulate but in practice this won't be many.