chat: fix content getting cut off/scroll stopping (#283233)
chat: fix content getting cut off/sticky scroll stopping

We never (apparently) correctly fired a content height change after the model got attached to the code blocks. We did did this is the CodeCompareBlockPart, just not the standard code block.

Also add an event on 'toolbar.onDidChangeMenuItems' -- don't this this ever caused issued but saw that when investigating.

Closes #276807