Fix editor edge clicking (#262964)
When clicking the 10 left-most pixels of the editor, the cursor was
moved to the beginning of the line. When clicking the 10 right-most
characters of the editor, the cursor was moved to the end of the line.
This is not correct in RTL. Clicking the left pixels should move the
cursor to the left. Clicking the right pixels should move the cursor to
the right.