node-pty@1.1.0-beta43
Brings in further improvements to handling of write backpressure. After
this we no longer throttle at all which means that for older versions of
bash at least on macOS interleaving is possible, but pasting and sending
should be near instantaneous for any reasonable amount of text. I
measured zsh injesting 5mb of data in ~14 seconds.

See:

- microsoft/node-pty#835
- microsoft/node-pty#837
- microsoft/node-pty#839

Part of #246204, #283056